+++ src/tools/index.ts
@@ -8,3 +8,4 @@ export * from './wiki.js';
 export * from './attachment.js';
 export * from './report.js';
 export * from './prd.js';
+export * from './ppt.js';